Module Pomap.Display_hasse_impl

module DefaultSpec : sig ... end

Default specification than can be used by the pretty-printer for Hasse-diagrams. Just include it into some module and override the defaults as required.

module Make (POMap : Pomap_intf.POMAP) (Spec : Display_hasse_intf.SPEC with type +'a node = 'a POMap.node) : Display_hasse_intf.DISPLAY_HASSE with type pomap = Spec.el POMap.pomap

Functor that generates a pretty-printer for Hasse-diagrams from a partially ordered map and a pretty-printer specification. See the Display_hasse_intf.DISPLAY_HASSE-interface for documentation.