Module Display_hasse_impl.Make

Functor that generates a pretty-printer for Hasse-diagrams from a partially ordered map and a pretty-printer specification. See the Display_hasse_intf.DISPLAY_HASSE-interface for documentation.

Parameters

module POMap : Pomap_intf.POMAP

module Spec : Display_hasse_intf.SPEC with type +'a node = 'a POMap.node

Signature

type pomap = Spec.el POMap.pomap

val fprintf : Stdlib.Format.formatter -> pomap -> unit

fprintf ppf pm prints partially ordered map pm to the pretty-printer ppf.

val printf : pomap -> unit

printf ppf pm prints partially ordered map pm to stdout.