Module type Display_hasse_intf.DISPLAY_HASSE

Interface for drawing Hasse-diagrams.

type pomap

val fprintf : Stdlib.Format.formatter -> pomap -> unit

fprintf ppf pm prints partially ordered map pm to the pretty-printer ppf.

val printf : pomap -> unit

printf ppf pm prints partially ordered map pm to stdout.